+++ CHANGELOG.md
@@ -6,6 +6,13 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 Unreleased
 **********
+* Upgraded to astroquery 0.4.9
+* Changed coordinateslookup.py to work with Pydantic even when the source doesn't have a velocity or redshift
+* For SIMBAD, the velocity or the redshift is set, depending on which is the master value
+* The redshift is now set if the source is found via NED
+* Test file renamed to test_coordinates.py so that tests will run
+* More tests were added for e.g. source in NED and not SIMBAD, as well as source not found at all
+* [BREAKING]: No source found returns a 404 error 
 
 0.4.0
 **********